requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

# RAG Pipeline
langchain==0.1.8
langchain-community==0.0.20
langchain-huggingface==0.0.3
faiss-cpu==1.7.4

# Document Processing
pypdf==4.0.1

# ML Models
transformers==4.37.2
accelerate==0.27.2
torch==2.1.2
sentence-transformers==2.2.2
safetensors==0.4.1

# NLP Dependencies
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl